Split dicom into two files and defer loading
